latest changes for new file structure (commonFunctions.js for both front-end and backend)
