+++ imports/startup/server/startup.js
@@ -2,10 +2,13 @@ import fs from 'fs-extra';
 import { Meteor } from 'meteor/meteor';
 import { logger } from 'meteor/votercircle:winston';
 import { moment } from 'meteor/momentjs:moment';
-import { LiveSites } from '../api/collections_server';
+import { LiveSites } from '../../api/collections_server';
 
 // Setting up directory in which this server expects incoming files (uses an environment variable)
 export const globalsite = LiveSites.findOne({ AQSID: `${process.env.aqsid}` });
+
+// reading ftps password from environment
+export const hnetsftp = process.env.hnetsftp;
 
 logger.info(`This instance is for AQSID ${process.env.aqsid} - ${globalsite.siteName}`);
