+++ server/main.js
@@ -1,4 +1,4 @@
 import '../imports/startup/startup-logger';
-import '../imports/startup/startup';
+import '../imports/startup/server/startup';
 import '../imports/methods/livefeedFunctions';
 import '../imports/cronjobs/cronJob_push';
